+++ python/aitemplate/frontend/nn/conv2d/conv2d_bias.py
@@ -69,7 +69,7 @@ class Conv2dBias(Conv2dBiasAct):
         padding=0,
         dilation=1,
         groups=1,
-        dtype="float16",
+        dtype="float32",
     ):
         super().__init__(
             "conv2d_bias",
